move settings to the header
