+++ src/components/Header/Header.tsx
@@ -1,9 +1,35 @@
-import { Avatar, Button, Flex, Group, Title } from "@mantine/core";
+import {
+  ActionIcon,
+  Avatar,
+  Button,
+  Flex,
+  Group,
+  MultiSelect,
+  Popover,
+  Space,
+  Text,
+  Title,
+  Tooltip,
+  useMantineColorScheme,
+} from "@mantine/core";
 import React from "react";
 import Icon from "../Icon/Icon";
 import { useUser } from "@auth0/nextjs-auth0/client";
+import { dayStore } from "@/lib/client/dayStore";
 
 const Header = () => {
+  const days = [
+    { label: "Su", value: "0" },
+    { label: "Mo", value: "1" },
+    { label: "Tu", value: "2" },
+    { label: "We", value: "3" },
+    { label: "Th", value: "4" },
+    { label: "Fr", value: "5" },
+    { label: "Sa", value: "6" },
+  ];
+  const day_store = dayStore();
+  const { toggleColorScheme } = useMantineColorScheme();
+
   const { user } = useUser();
   const isLoggedIn = Boolean(user);
   return (
@@ -15,29 +41,77 @@ const Header = () => {
         >
           Schedule Builder
         </Title>
-        {isLoggedIn ? (
-          <Group>
+        <Group>
+          {isLoggedIn ? (
+            <>
+              <Button
+                component={"a"}
+                href={"/api/auth/logout"}
+                rightSection={<Icon>logout</Icon>}
+              >
+                Logout
+              </Button>
+
+              <Avatar src={user?.picture} alt={user?.name ?? ""} />
+            </>
+          ) : (
             <Button
               component={"a"}
-              href={"/api/auth/logout"}
-              rightSection={<Icon>logout</Icon>}
+              href={"/api/auth/login"}
+              rightSection={<Icon>login</Icon>}
             >
-              Logout
+              Login
             </Button>
-            {/* <Button onClick={uploadPlan} rightSection={<Icon>upload</Icon>}>
-              Upload Plan
-            </Button> */}
-            <Avatar src={user?.picture} alt={user?.name ?? ""} />
-          </Group>
-        ) : (
-          <Button
-            component={"a"}
-            href={"/api/auth/login"}
-            rightSection={<Icon>login</Icon>}
-          >
-            Login
-          </Button>
-        )}
+          )}
+
+          <Popover width={350} position="bottom" withArrow shadow="md">
+            <Popover.Target>
+              <ActionIcon
+                variant="light"
+                p={"md"}
+                radius="xl"
+                aria-label="Settings"
+              >
+                <Icon>settings</Icon>
+              </ActionIcon>
+            </Popover.Target>
+            <Popover.Dropdown>
+              <Group>
+                <Group m="sm">
+                  <Tooltip label="Toggle color scheme">
+                    <ActionIcon
+                      className="m-1"
+                      variant="light"
+                      onClick={toggleColorScheme}
+                    >
+                      <Icon>
+                        <p className="dark:hidden">light_mode</p>
+                        <p className="hidden dark:block">dark_mode</p>
+                      </Icon>
+                    </ActionIcon>
+                  </Tooltip>
+                  <Text size="md">Color Scheme</Text>
+                </Group>
+                <Space h="md" />
+                <Text size="md" ta={"center"}>
+                  Days of the Week
+                </Text>
+                <MultiSelect
+                  comboboxProps={{ withinPortal: false }}
+                  mx="md"
+                  label="Hidden days"
+                  placeholder="Select hidden days"
+                  data={days}
+                  maxValues={6}
+                  onChange={(values) => {
+                    day_store.setDays(values.map((day) => parseInt(day)));
+                  }}
+                  value={day_store.days.map((day) => day.toString())}
+                />
+              </Group>
+            </Popover.Dropdown>
+          </Popover>
+        </Group>
       </Flex>
     </>
   );
